updated tests, added new sdk functions, added started the migration of keeping all mocks in _mocks.go files
